Refactored the button components
