+++ cccm-frontend/src/components/common/templateButtonGroupSubmit.json
@@ -1,0 +1,61 @@
+{
+  "display": "form",
+  "components": [
+    {
+      "label": "Columns",
+      "customClass": "d-flex justify-content-between p-1",
+      "columns": [
+          {
+              "components": [
+                {
+                  "customClass": "text-left pl-4",
+                  "label": "Save and Continue",
+                  "action": "event",
+                  "showValidations": false,
+                  "disableOnInvalid": true,
+                  "tableView": true,
+                  "key": "save_continue",
+                  "type": "button",
+                  "input": false,
+                  "event": "evt_save",
+                  "theme": "primary"
+                }
+              ],
+              "width": 6,
+              "offset": 0,
+              "push": 0,
+              "pull": 0,
+              "size": "md",
+              "currentWidth": 6
+          },
+          {
+              "components": [
+                {
+                  "customClass": "text-right mr-2",
+                  "label": "Delete Form",
+                  "action": "event",
+                  "showValidations": false,
+                  "disableOnInvalid": true,
+                  "tableView": true,
+                  "key": "add_cancel",
+                  "type": "button",
+                  "input": false,
+                  "event": "evt_cancel",
+                  "theme": "danger"
+                }
+              ],
+              "width": 6,
+              "offset": 0,
+              "push": 0,
+              "pull": 0,
+              "size": "md",
+              "currentWidth": 6
+          }
+      ],
+      "key": "columns",
+      "type": "columns",
+      "input": false,
+      "tableView": false
+    }
+  ]
+}
